update item page on react app to display item data

+++ src/main/java/com/example/springboot/ItemServiceImplementation.java
@@ -30,7 +30,7 @@ public class ItemServiceImplementation {
             ResultSet rs = stmt.executeQuery();
             while(rs.next())
             {
-                Items item = new Items(rs.getInt(1), rs.getString(2), rs.getString(3), rs.getDouble(4), rs.getString(5));
+                Items item = new Items(rs.getInt(1), rs.getString(2), rs.getString(3), rs.getFloat(4), rs.getString(5));
                 itemList.add(item);
             }
         }
